site: remove circ dependency
