+++ site/src/modules/resources/AgentRow.tsx
@@ -8,20 +8,12 @@ import type {
 	Workspace,
 	WorkspaceAgent,
 	WorkspaceAgentMetadata,
-	WorkspaceApp,
 } from "api/typesGenerated";
 import { isAxiosError } from "axios";
 import { Button } from "components/Button/Button";
 import { DropdownArrow } from "components/DropdownArrow/DropdownArrow";
-import {
-	DropdownMenu,
-	DropdownMenuContent,
-	DropdownMenuItem,
-	DropdownMenuTrigger,
-} from "components/DropdownMenu/DropdownMenu";
 import { Stack } from "components/Stack/Stack";
 import { useProxy } from "contexts/ProxyContext";
-import { Folder } from "lucide-react";
 import { useFeatureVisibility } from "modules/dashboard/useFeatureVisibility";
 import { AppStatuses } from "pages/WorkspacePage/AppStatuses";
 import {
@@ -36,7 +28,7 @@ import {
 import { useQuery } from "react-query";
 import AutoSizer from "react-virtualized-auto-sizer";
 import type { FixedSizeList as List, ListOnScrollProps } from "react-window";
-import { AgentButton } from "./AgentButton";
+import { Apps, organizeAgentApps } from "./AgentApps/AgentApps";
 import { AgentDevcontainerCard } from "./AgentDevcontainerCard";
 import { AgentLatency } from "./AgentLatency";
 import { AGENT_LOG_LINE_HEIGHT } from "./AgentLogs/AgentLogLine";
@@ -44,7 +36,6 @@ import { AgentLogs } from "./AgentLogs/AgentLogs";
 import { AgentMetadata } from "./AgentMetadata";
 import { AgentStatus } from "./AgentStatus";
 import { AgentVersion } from "./AgentVersion";
-import { AppLink } from "./AppLink/AppLink";
 import { DownloadAgentLogsButton } from "./DownloadAgentLogsButton";
 import { PortForwardButton } from "./PortForwardButton";
 import { AgentSSHButton } from "./SSHButton/SSHButton";
@@ -351,93 +342,6 @@ export const AgentRow: FC<AgentRowProps> = ({
 				</section>
 			)}
 		</Stack>
-	);
-};
-
-export type AppSection = {
-	/**
-	 * If there is no `group`, just render all of the apps inline. If there is a
-	 * group name, show them all in a dropdown.
-	 */
-	group?: string;
-
-	apps: WorkspaceApp[];
-};
-
-/**
- * organizeAgentApps returns an ordering of agent apps that accounts for
- * grouping. When we receive the list of apps from the backend, they have
- * already been "ordered" by their `order` attribute, but we are not given that
- * value. We must be careful to preserve that ordering, while also properly
- * grouping together all apps of any given group.
- *
- * The position of the group overall is determined by the `order` position of
- * the first app in the group. There may be several sections returned without
- * a group name, to allow placing grouped apps in between non-grouped apps. Not
- * every ungrouped section is expected to have a group in between, to make the
- * algorithm a little simpler to implement.
- */
-export function organizeAgentApps(apps: readonly WorkspaceApp[]): AppSection[] {
-	let currentSection: AppSection | undefined = undefined;
-	const appGroups: AppSection[] = [];
-	const groupsByName = new Map<string, AppSection>();
-
-	for (const app of apps) {
-		if (app.hidden) {
-			continue;
-		}
-
-		if (!currentSection || app.group !== currentSection.group) {
-			const existingSection = groupsByName.get(app.group!);
-			if (existingSection) {
-				currentSection = existingSection;
-			} else {
-				currentSection = {
-					group: app.group,
-					apps: [],
-				};
-				appGroups.push(currentSection);
-				if (app.group) {
-					groupsByName.set(app.group, currentSection);
-				}
-			}
-		}
-
-		currentSection.apps.push(app);
-	}
-
-	return appGroups;
-}
-
-export type AppsProps = {
-	section: AppSection;
-	agent: WorkspaceAgent;
-	workspace: Workspace;
-};
-
-export const Apps: FC<AppsProps> = ({ section, agent, workspace }) => {
-	return section.group ? (
-		<DropdownMenu>
-			<DropdownMenuTrigger asChild>
-				<AgentButton>
-					<Folder />
-					{section.group}
-				</AgentButton>
-			</DropdownMenuTrigger>
-			<DropdownMenuContent align="start">
-				{section.apps.map((app) => (
-					<DropdownMenuItem key={app.slug}>
-						<AppLink grouped app={app} agent={agent} workspace={workspace} />
-					</DropdownMenuItem>
-				))}
-			</DropdownMenuContent>
-		</DropdownMenu>
-	) : (
-		<>
-			{section.apps.map((app) => (
-				<AppLink key={app.slug} app={app} agent={agent} workspace={workspace} />
-			))}
-		</>
 	);
 };
